backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
anthropic>=0.42.0
python-dotenv>=1.0.0
numpy>=1.26.0

# Optional: speaker embedding (improves speaker ID accuracy)
# Uncomment after testing basic flow works
# resemblyzer>=0.1.3
# librosa>=0.10.0
